backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
websockets==15.0.1
python-multipart==0.0.20
pydantic==2.11.3
email-validator==2.2.0
httpx==0.28.1
qrcode[pil]==8.2
itsdangerous==2.2.0
pymongo==4.12.0
python-dotenv==1.0.1
